+++ src/Sensorario/ValueObject/Validators/MandatoryProperties.php
@@ -20,6 +20,18 @@ final class MandatoryProperties implements Validator
     public static function check(ValueObject $valueObject)
     {
         foreach ($valueObject->mandatory() as $key => $value) {
+            if (isset($value['when'])) {
+                $propertyName = $value['when']['property'];
+                $propertyValue = $value['when']['value'];
+                if ($valueObject->get($propertyName) === $propertyValue) {
+                    throw new RuntimeException(
+                        'When property `' . $propertyName . '` has value '
+                        . '`' . $propertyValue . '` also `' . $key . '` '
+                        . 'is mandatory'
+                    );
+                }
+            }
+
             if (is_numeric($key) && $valueObject->hasNotProperty($value)) {
                 if (!isset($valueObject->defaults()[$value])) {
                     throw new RuntimeException(
@@ -29,7 +41,12 @@ final class MandatoryProperties implements Validator
                 }
             }
 
-            if (!is_numeric($key) && $valueObject->hasProperty($value['if_present'])) {
+            $mandatoryIfPresentAnotherValue = false;
+            if (isset($value['if_present'])) {
+                $mandatoryIfPresentAnotherValue = $valueObject->hasProperty($value['if_present']);
+            }
+
+            if (!is_numeric($key) && $mandatoryIfPresentAnotherValue) {
                 if ($valueObject->hasNotProperty($key)) {
                     throw new RuntimeException(
                         "Property `" . get_class($valueObject)